+++ includes/Thread/ICondVar.hpp
@@ -5,7 +5,7 @@
 // Login   <porres_m@epitech.net>
 // 
 // Started on  Mon Apr 20 15:18:12 2015 Martin Porrès
-// Last update Sun May 24 17:14:18 2015 Martin Porrès
+// Last update Mon May 25 15:40:10 2015 Martin Porrès
 //
 
 #ifndef		_ICONDVAR_HPP_
@@ -16,7 +16,7 @@ class		ICondVar
 public:
   virtual ~ICondVar(void) {};
   virtual void	wait(void) = 0;
-  virtual void	timedwait(double time) = 0;
+  virtual void	timedwait(long time) = 0;
   virtual void	signal(void) = 0;
   virtual void	broadcast(void) = 0;
 };
